+++ Potraviny/src/main/java/Zelenina/Mrkva.java
@@ -17,7 +17,7 @@ public class Mrkva extends Ovocie_Zelenina {
     private Farba farba;
     private boolean korenovaZelenina;
 
-    public Mrkva(Farba farba, boolean korenovaZelenina, String nazov, double hmotnost, int cena, boolean sladke, boolean supka, boolean kysele) {
+    public Mrkva(Farba farba, boolean korenovaZelenina, String nazov, int hmotnost, int cena, boolean sladke, boolean supka, boolean kysele) {
         super(nazov, hmotnost, cena, sladke, supka, kysele);
         this.farba = farba;
         this.korenovaZelenina = korenovaZelenina;
@@ -49,11 +49,11 @@ public class Mrkva extends Ovocie_Zelenina {
         this.nazov = nazov;
     }
 
-    public double getHmotnost() {
+    public int getHmotnost() {
         return hmotnost;
     }
 
-    public void setHmotnost(double hmotnost) {
+    public void setHmotnost(int hmotnost) {
         this.hmotnost = hmotnost;
     }
